Move delimiter utilities to common module
- Moved Delimiter and OneOfDelimiters to common/interim-input-utils
- Removed duplicate definitions from T2 and T3 implementations
